loads big script when uploaded 3x preformance.
loads it using sveltekits module.
