+++ .timetracker
@@ -1,0 +1,1 @@
+{"total":426,"sessions":[{"begin":"2024-11-01T19:25:07+01:00","end":"2024-11-01T19:29:37+01:00","duration":269},{"begin":"2024-11-01T19:57:12+01:00","end":"2024-11-01T19:59:49+01:00","duration":157}]}
