+++ src/routes/statistics/+page.js
@@ -104,7 +104,6 @@ if (browser) {
 		leastVotes.textContent = generalData['numVotes']['lowest']['votes'];
 		mostVotesTitle.textContent = generalData['numVotes']['highest']['title'];
 		leastVotesTitle.textContent = generalData['numVotes']['lowest']['title'];
-		console.log(movies['perMonth'], generalData);
 		loadCharts();
 	}
